Fix bug that blocks creating/launching instances across multiple ECS profiles and subnets
If you have two ECS cluster profiles which have different subnet configurations, it is possible for the plugin to get stuck and be unable to create instances. This is because when instances/containers are searched for on EC2, they are not filtered by the `cluster-name`, so you may get returned instances for a different cluster profile than the one you are trying to create. That may be intentional, as the goal here is to launch an instance in the best subnet, and cluster profiles may be sharing subnets.

If you end up with state
cluster-profile-A --> instance-A --> subnet-A

and you then need to launch an instance in

cluster-profile-B in one of (subnet-B, subnet-C)

... the previous code fails to handle it, because is finds an instance, but then filters by relevant subnets, and then doesn't handle the case where there are no subnets after counting by index.