Support Docker 25 and AL2023 by removing instance store Docker volume initialization 'magic'
The plugin previously had some userdata/cloud-init magic to utilise any special NVMe instance stores for Docker
storage by creating a new logical volume group and restarting Docker on this volume. This would work "magically" on 'd'
instances (e.g c6id). Additionally changes to ECS and cloud-init makes the previous approach to installing various packages
not work any longer due to changes within ECS/AL2023, so the volume tools need additional work to get in the right state.

This is broken on Docker 25 and thus AL2023 because it relied on the docker device-mapper storage driver which is
no longer available, and there was no way to opt out of it. In any case, this no longer seems necessary, as for the
purposes of this image it seems regular EBS GP3 style volumes should probably be sufficient for use within the GoCD
build agent containers being run.

If there is a need to add this back, we'll add this back based on user feedback and assistance, but right now this seems to be doing more harm than good, so will remove the support.